Split into main.rs and lib.rs
